# swagger_bid.json
{
  "components": {
    "parameters": {
      "authHeader": {
        "description": "Токен для авторизации",
        "in": "header",
        "name": "Authorization",
        "required": true,
        "schema": {
          "example": "Bearer ACCESS_TOKEN",
          "type": "string"
        }
      }
    },
    "responses": {
      "badRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        },
        "description": "Bad request"
      },
...
      "internalError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        },
        "description": "Internal Error"
      },
...
      "unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        },
        "description": "Unauthorized"
      }
    },
    "schemas": {
...
      "auto": {
        "nullable": true,
        "properties": {
          "budgetPenny": {
            "description": "Текущий бюджет (в копейках)",
            "nullable": true,
            "type": "integer"
          },
          "budgetType": {
            "description": "Тип бюджета (\"1d\" - дневной | \"7d\" - недельный | \"30d\" - месячный)",
            "nullable": true,
            "type": "string"
          },
          "dailyBudget": {
            "$ref": "#/components/schemas/budget"
          },
          "maxBudgetPenny": {
            "description": "Максимальный бюджет (в копейках)",
            "type": "integer"
          },
          "minBudgetPenny": {
            "description": "Минимальный бюджет (в копейках)",
            "type": "integer"
          },
          "monthlyBudget": {
            "$ref": "#/components/schemas/budget"
          },
          "weeklyBudget": {
            "$ref": "#/components/schemas/budget"
          }
        },
        "required": [
          "minBudgetPenny",
          "maxBudgetPenny",
          "dailyBudget",
          "weeklyBudget",
          "monthlyBudget"
        ],
        "type": "object"
      },
...
      "bid": {
        "properties": {
          "compare": {
            "description": "Количество процентных пунктов. (На сколько выше в сравнении с конкурентами)",
            "type": "integer"
          },
          "maxForecast": {
            "description": "Максимальное кол-во ЦД за ставку",
            "type": "integer"
          },
          "minForecast": {
            "description": "Минимальное кол-во ЦД за ставку",
            "type": "integer"
          },
          "valuePenny": {
            "description": "Ставка (в копейках)",
            "type": "integer"
          }
        },
        "required": [
          "valuePenny",
          "minForecast",
          "maxForecast",
          "compare"
        ],
        "type": "object"
      },
      "budget": {
        "properties": {
          "budgets": {
            "items": {
              "$ref": "#/components/schemas/budgetValue"
            },
            "maxItems": 200,
            "minItems": 1,
            "type": "array"
          },
          "maxLimitPenny": {
            "description": "Максимальный лимит (в копейках)",
            "type": "integer"
          },
          "minLimitPenny": {
            "description": "Минимальный лимит (в копейках)",
            "type": "integer"
          },
          "recBudgetPenny": {
            "description": "Рекомендованный лимит (в копейках)",
            "type": "integer"
          }
        },
        "required": [
          "budgets",
          "minLimitPenny",
          "maxLimitPenny",
          "recBudgetPenny"
        ],
        "type": "object"
      },
      "budgetValue": {
        "properties": {
          "compare": {
            "description": "Количество процентных пунктов. (На сколько выше в сравнении с конкурентами)",
            "type": "integer"
          },
          "maxForecast": {
            "description": "Максимальное кол-во ЦД за ставку",
            "type": "integer"
          },
          "minForecast": {
            "description": "Минимальное кол-во ЦД за ставку",
            "type": "integer"
          },
          "valuePenny": {
            "description": "Бюджет (в копейках)",
            "type": "integer"
          }
        },
        "required": [
          "valuePenny",
          "minForecast",
          "maxForecast",
          "compare"
        ],
        "type": "object"
      },
      "error": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "description": "Текст ошибки",
            "example": "Что-то пошло не так",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
      "getBidsOut": {
        "properties": {
          "actionTypeID": {
            "description": "Тип события (1 - звонок | 5 - пакет кликов | 7 - мессенджер, передача контакта в чате)",
            "type": "integer"
          },
          "auto": {
            "$ref": "#/components/schemas/auto"
          },
          "manual": {
            "$ref": "#/components/schemas/manual"
          },
          "selectedType": {
            "description": "Выбранный тип продвижения (\"manual\" - ручное продвижение | \"auto\" - автоматическое продвижение)",
            "type": "string"
          }
        },
        "required": [
          "actionTypeID",
          "selectedType"
        ],
        "type": "object"
      },
...
      "manual": {
        "nullable": true,
        "properties": {
          "bidPenny": {
            "description": "Текущая ставка (в копейках)",
            "nullable": true,
            "type": "integer"
          },
          "bids": {
            "items": {
              "$ref": "#/components/schemas/bid"
            },
            "maxItems": 200,
            "minItems": 1,
            "type": "array"
          },
          "limitPenny": {
            "description": "Дневной лимит (в копейках)",
            "nullable": true,
            "type": "integer"
          },
          "maxBidPenny": {
            "description": "Максимальная ставка (в копейках)",
            "type": "integer"
          },
          "maxLimitPenny": {
            "description": "Максимальный дневной лимит (в копейках)",
            "type": "integer"
          },
          "minBidPenny": {
            "description": "Минимальная ставка (в копейках)",
            "type": "integer"
          },
          "minLimitPenny": {
            "description": "Минимальный дневной лимит (в копейках)",
            "type": "integer"
          },
          "recBidPenny": {
            "description": "Рекомендуемая ставка (в копейках)",
            "type": "integer"
          }
        },
        "required": [
          "recBidPenny",
          "minBidPenny",
          "maxBidPenny",
          "minLimitPenny",
          "maxLimitPenny",
          "bids"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AuthorizationCode": {
        "description": "Это API использует OAuth 2 с механизмом authorization_code. Используйте его для доступа к данным других пользователей при разработке стороннего приложения. [Подробнее](#tag/ApplicationAccess)",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://avito.ru/oauth",
            "scopes": {
              "cpxpromo:edit": "Редактирование ставок",
              "cpxpromo:read": "Получение ставок"
            },
            "tokenUrl": "https://api.avito.ru/token"
          }
        },
        "type": "oauth2"
      },
      "ClientCredentials": {
        "description": "Это API использует OAuth 2 с механизмом client_credentials. Используйте его для доступа к возможностям своей личной учетной записи. [Подробнее](#tag/Access)",
        "flows": {
          "clientCredentials": {
            "scopes": {

            },
            "tokenUrl": "https://api.avito.ru/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Методы для работы с сервисом Cpxpromo.\n\n**Авито API для бизнеса предоставляется согласно [Условиям использования](https://www.avito.ru/legal/pro_tools/public-api).**\n",
    "title": "Cpxpromo",
    "version": "1"
  },
  "openapi": "3.0.0",
  "paths": {
    "/cpxpromo/1/getBids/{itemId}": {
      "get": {
        "description": "Метод позволяет получить детализированную информацию о действующих и доступных ценах целевого действия (в копейках), бюджетах (в копейках) и преимуществах перед конкурентами.\n",
        "operationId": "getBids",
        "parameters": [
          {
            "$ref": "#/components/parameters/authHeader"
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getBidsOut"
                }
              }
            },
            "description": "Ok"
          },
          "400": {
            "$ref": "#/components/responses/badRequest"
          },
          "401": {
            "$ref": "#/components/responses/unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/internalError"
          }
        },
        "security": [
          {
            "AuthorizationCode": [
              "cpxpromo:read"
            ]
          },
          {
            "ClientCredentials": []
          }
        ],
        "summary": "Получение детализированной информации о действующих и доступных ставках и бюджетах",
        "x-rate-limiter": {
          "default": 1000
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.avito.ru/"
    }
  ],
...
}